triangle_lab
Task:
Write a program that creates and initializes 2 objects of the triangle class in the Cartesian coordinate system. Provide the ability to rotate a triangle around the origin by a given angle, calculate the area and perimeter using class methods. The use of constructors and destructors is mandatory. Class instances must be created dynamically.